+++ src/io/github/sspanak/tt9/AbsSymDialog.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.traditionalt9;
+package io.github.sspanak.tt9;
 
 import android.app.Dialog;
 import android.content.Context;
