+++ src/io/github/sspanak/tt9/CandidateView.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.traditionalt9;
+package io.github.sspanak.tt9;
 
 import android.content.Context;
 import android.content.res.Resources;
