+++ src/io/github/sspanak/tt9/DBUpdateService.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.traditionalt9;
+package io.github.sspanak.tt9;
 
 import android.app.IntentService;
 import android.app.Notification;
