+++ src/io/github/sspanak/tt9/KeyMap.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.traditionalt9;
+package io.github.sspanak.tt9;
 
 import android.os.Environment;
 import android.util.Log;
