+++ src/io/github/sspanak/tt9/LangHelper.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.traditionalt9;
+package io.github.sspanak.tt9;
 
 import java.util.HashMap;
 import java.util.Locale;
