+++ src/io/github/sspanak/tt9/settings/SettingList.java
@@ -1,13 +1,13 @@
-package io.github.sspanak.traditionalt9.settings;
+package io.github.sspanak.tt9.settings;
 
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.util.AttributeSet;
 
-import io.github.sspanak.traditionalt9.R;
-import io.github.sspanak.traditionalt9.T9DB;
-import io.github.sspanak.traditionalt9.T9DB.DBSettings.SETTING;
+import io.github.sspanak.tt9.R;
+import io.github.sspanak.tt9.T9DB;
+import io.github.sspanak.tt9.T9DB.DBSettings.SETTING;
 
 public class SettingList extends Setting {
 	String[] entries;
